Move test Dart code for json_serializable_test to a dart file
Makes it a lot easier to maintain this content
Using //ignore comments for intentionally invalid code
