+++ json_serializable/test/src/json_serializable_test_input.dart
@@ -1,0 +1,149 @@
+import 'package:json_annotation/json_annotation.dart';
+
+@JsonSerializable()
+const theAnswer = 42;
+
+@JsonSerializable()
+void annotatedMethod() => null;
+
+@JsonSerializable()
+class Person {
+  String firstName, lastName;
+  @JsonKey(name: 'h')
+  int height;
+  DateTime dateOfBirth;
+  dynamic dynamicType;
+  var varType;
+  List<int> listOfInts;
+}
+
+@JsonSerializable()
+class Order {
+  final String firstName, lastName;
+  int height;
+  DateTime dateOfBirth;
+
+  Order(this.height, String firstName, [this.lastName])
+      : this.firstName = firstName;
+}
+
+@JsonSerializable()
+class FinalFields {
+  final int a;
+  int get b => 4;
+
+  FinalFields(this.a);
+}
+
+@JsonSerializable()
+class FromJsonOptionalParameters {
+  final ChildWithFromJson child;
+
+  FromJsonOptionalParameters(this.child);
+}
+
+class ChildWithFromJson {
+  ChildWithFromJson.fromJson(json, {initValue: false});
+}
+
+@JsonSerializable()
+class ParentObject {
+  int number;
+  String str;
+  ChildObject child;
+}
+
+@JsonSerializable()
+class ChildObject {
+  int number;
+  String str;
+
+  factory ChildObject.fromJson(json) => null;
+}
+
+@JsonSerializable()
+class ParentObjectWithChildren {
+  int number;
+  String str;
+  List<ChildObject> children;
+}
+
+@JsonSerializable()
+class ParentObjectWithDynamicChildren {
+  int number;
+  String str;
+  List<dynamic> children;
+}
+
+@JsonSerializable()
+class UnknownCtorParamType {
+  int number;
+
+  // ignore: undefined_class, field_initializer_not_assignable
+  UnknownCtorParamType(Bob number) : this.number = number;
+}
+
+@JsonSerializable()
+class UnknownFieldType {
+  // ignore: undefined_class
+  Bob number;
+}
+
+@JsonSerializable(createFactory: false)
+class NoSerializeFieldType {
+  Stopwatch watch;
+}
+
+@JsonSerializable(createToJson: false)
+class NoDeserializeFieldType {
+  Stopwatch watch;
+}
+
+@JsonSerializable(createFactory: false)
+class NoSerializeBadKey {
+  Map<int, DateTime> intDateTimeMap;
+}
+
+@JsonSerializable(createToJson: false)
+class NoDeserializeBadKey {
+  Map<int, DateTime> intDateTimeMap;
+}
+
+@JsonSerializable(createFactory: false)
+class IncludeIfNullAll {
+  @JsonKey(includeIfNull: true)
+  int number;
+  String str;
+}
+
+@JsonSerializable(createFactory: false, includeIfNull: false)
+class IncludeIfNullOverride {
+  @JsonKey(includeIfNull: true)
+  int number;
+  String str;
+}
+
+// https://github.com/dart-lang/json_serializable/issues/7 regression
+@JsonSerializable()
+class NoCtorClass {
+  final int member;
+
+  factory NoCtorClass.fromJson(Map<String, dynamic> json) => null;
+}
+
+@JsonSerializable(createFactory: false)
+class KeyDupesField {
+  @JsonKey(name: 'str')
+  int number;
+
+  String str;
+}
+
+@JsonSerializable(createFactory: false)
+class DupeKeys {
+  @JsonKey(name: 'a')
+  int number;
+
+  @JsonKey(name: 'a')
+  String str;
+}
